+++ src/OffspringPortal/PrefabNames.cs
@@ -1,0 +1,7 @@
+namespace OffspringPortal;
+
+public static class PrefabNames
+{
+    public const string OffspringPortal = "offspring_portal";
+    public const string VanillaPortalWood = "portal_wood";
+}